+++ monopoly/src/router/index.ts
@@ -21,11 +21,6 @@ const routes: RouteRecordRaw[] = [
     path: '/game',
     name: 'game',
     component: () => import('@/views/GameView.vue')
-  },
-  {
-    path: '/result',
-    name: 'result',
-    component: () => import('@/views/ResultView.vue')
   }
 ]